Initial Python and Fortran Files
Uploading several of the initial python and fortran scripts that i've written to process FEpX data. It has a lot of the same features as the matlab based code, OdfPF, but it's still at parity with it yet. It also has several features that the matlab based code does not.