Bug Fixes and Updated Code
Finally fixed some long standing bugs that existed in the Misori
module. I’ve also added angle-axis vector as an output to
misorientationStats.

Fixed some small issues in FePX_Data_and_Mesh.py

Removed a lot of small files that shouldn’t have been included in the
original packaging of this.

Added some new code to the PythonFortranCode matrixMath module. This
code will be used in finding the barycentric coordinates of a
tetrahedron and the wts that go along with that. This will be used
heavily in getting pyFEpX up to being able to find ODFs using FEs. In
other words these functions will be necessary in making pyFEpX on par
with OdfPf in MATLAB.