Added a scalar gradient function
Added a method that can calculate from a set of nodal scalar values on
a finite element mesh the gradient of those scalar values in an element.

This method could also be easily extended to find the gradient of a
vector field in a finite element mesh as well.

Also made some small bug fixes as well to Misori module and the fortran
code matrixMath.