Updated code for the LOFEM Refactor
Added a job script to create parts of a LOFEM job.

Added an example script to process LOFEM data with the new LOFEM
refactored structure.

Updated Several of the mesh and data file readers that needed to be
expanded for the new LOFEM refactor.

Other updates are related to the graph and connected component creation
routines that are needed for studying network formation in a simulation.