Dislocation Density python function and helpers
Created a purely python based version to calculate the dislocation
density using the L2 method from Arsenlis 1999 paper. Added a function
to calculate the Nye tensor. Also, added a function to calculate a
gradient of a vector field. Then in the local gradient shape functions
function I have it outputting the determinant for the element at that
gauss point as well.