Bug fix and python version of dislocation density
Found a bug in misorientation calculations that carried over from the
matlab code. It appears to really only have an affect when large
misorientations occur in a grain. This is due to the original matlab
code calculating an average theta value for each axial vector instead
of the individual value. I missed this during the original port of the
code. However, when small misorientations occur this average value is
fairly close to the actual value, and so it doesn’t change the answer
by any noticeable amount.  The new code properly addresses this issue.

Next, a version of the dislocation density code that just relies on
python code is available. It should be used over the python script that
uses Fortran code until I can properly find out what in the Fortran
code is causing it to diverge from the correct answer. My current
thought is this is an issue due to variable being sent in with a C
memory allocation instead of Fortran. When I have time I’ll need to
chase down the exact issue. In the mean time, one should use the python
version, since it is correct and only slightly slower.