Fixed Rotation bugs
Fixed a bug in the Rotations module related to QuatOfRMat and how numpy doesn't realize that a (n) and a (n,1) size vector are the exact same thing.
Next simplified some of the QuatOfRod function to use a norm function instead of doing it by hand.
Fixed QuatMean where I left off the eig function name.
Fixed a small issue in QuatOfAngleAxis where around pi or -pi rotations can cause widely different answers from what the original value should be. It has now been fixed such that this no longer appears to be a problem. Though, it still might be depending on how sensitive it is around that value.