Example file for finding grain boundary elements and nodes
Created an example showing how to go about finding all of the nodes and
elements that lie on a grain boundary. It also goes grain by grain and
puts all of the nodes and elements that lay on the GB for that grain
into a set. This set is then appended to a growing list that is in
order of the grain number.