fixStrain related bug fixes
Realized I'd upated how fixStrain outputs stuff, and I hadn't updated the relevant scripts that use it. It's because of this that I also updated the deformation metric functions.

Finally, I removed an old script that had been replaced with one that's more up to date and actually commented on.